Add inliner support for alloc_ref [stack].
When the callee loads from an alloc_stack we try to inline it. Do the
same for alloc_ref [stack].

It probably makes sense to do this whenever the callee accesses
something allocated in the caller, regardless of whether it's on the
stack. But this is more conservative for now.