Remove InoutAliasingAssumption, improving EscapeAnalysis.
Simplifies isNotAliasingArgument and pointsToLocalObject APIs.

Rename to isExclusiveArgument to avoid double-negative bugs.

The motivation is to simplify the APIs, but a side effect is that
EscapeAnlysis will no longer consider @inout arguments to be aliasing
with other indirect arguments or globals. This is valid since
exclusivity has been fully enforced for at least a major release cycle.