SILVerifier: witness_method lookup type must match apply's Self.
This enforces a new rule:

  witness_method $Conformance, #Protocol.member : <Self where Self : Protocol> (inout Self) -> () -> ()

can only be applied as:

 apply <Conformance Protocol>() : $@convention(witness_method: Protocol) <τ_0_0 where τ_0_0 : Protocol> (@inout τ_0_0) -> ()

Previously, we allowed refining the witness method without changing the apply's function type, so this would be valid:

  witness_method $Concrete, #Protocol.member : <Self where Self : Protocol> (inout Self) -> () -> ()
  apply <Conformance Protocol>() : $@convention(witness_method: Protocol) <τ_0_0 where τ_0_0 : Protocol> (@inout τ_0_0) -> ()

However, this breaks an assumption made by the Devirtualizer that the
witness_method can be directly replaced by its concrete
function_ref. When the Devirtualizer substitues the function_ref, it
rewrites the apply's function type. If the witness_method's lookup
type does not match the apply's Self type, then substituting the
function_ref will create a mismatch between the apply's substitutions
and Self argument's type constraints:

  function_ref @function_name : $@convention(witness_method: Protocol) <τ_0_0 where τ_0_0 : Concrete> (@inout τ_0_0) -> ()
  apply <Conformance Protocol>(%11) : $@convention(witness_method: Protocol) <τ_0_0 where τ_0_0 : Concrete> (@inout τ_0_0) -> ()

This is a problem because Conformance is not constrained to be Concrete.

Previously, this error would never show up unless the apply site
were later inlined and the caller's Conformance-typed argument did not
meet the same requirements as the callee's Concrete
constraint. e.g. Concrete might be a loadable type while Conformance
is not a loadable type.

An alternative might be to make the Devirtualizer resolve type
mismatches, but the immediate priority is to catch errors early, and
it's not clear that it is ever a good idea to refine the
witness_method without refining the apply's Self arguments too.