SILInstructionWorklist - add createCallbacks and madeChange.
Move the callback and madeChange handling into the worklist where it
belongs. This logic was inconsistently scattered across all the
clients, leading to bugs.

The SILInstructionWorklist API can gradually be simplified until it no
longer does anything except handle callbacks and record changes. The
rest of the API belongs in InstructionDeleter.