Cleanup PrunedLiveBlocks
Remove the bitfield from PrunedLiveBlocks. Use BasicBlockBitfield
instead. This (re)specializes the data structure for scalar liveness
and drastically simplifies the implementation.

This utility is fundamental to all ownership utilities. It will be on
the critical path in many areas of the compiler, including at
-Onone. It needs to be minimal and as easy as possible for compiler
engineers to understand, investigate, and debug.