Revert "Fix operand ownership of mark_dependence [nonescaping] of address values"
Revert recent change to mark_dependence ownership.

This was causing OSSA verification failures.

When the base is a non-address value, the incoming operand ownership
is always a Borrow of the base operand, which may be an owned
value. The mark_dep introduces a new borrow scope.

If the mark_dep result is an address, the client still needs to be
handle it like an interior pointer. But that's different from the
operand ownership.

This reverts commit 79b649854b0a39b5aa0f79d826553004bd460c92.