LifetimeDependenceDefUseWalker: force liveness at a destroy_addr.
When a dependent value is stored in a local stack location, guarantee that the
dependency source is extended over the local's destroy_addr, not just the last
reachable use.

This has no practical impact on most non-Escapable values today which are
trivial, but will be needed for nontrivial non-Escapable values.