TypeLowering: remove special IsLexical treat ment for non-Copyable
TypeLowering tracks whether a type has the @eagerMove attribute. This is
unrelated to whether a type is non-Copyable. non-Copyable types have strict
lifetimes. This is not the same thing as "lexical". Remove the special handling
in TypeLowering to eliminate a source of confusion and allow non-Copyable
lifetimes to be optimized based on other factors, such as the absense of a
deinit.